python_tests: Avoid creating .pylint.d directory
pylint creates .pylint.d/setup1.stats when run. This file will end up in
the Debian package when using pybuild. This file does not belong in the
produced package and if two packages uses this snippet, they will ship
the file in the same path and conflict with each other.

Disable the persistence of pylint to prevent it to create this file.
Then PYLINTHOME can point to a non-existing directory again.